feat; 지도 컴포넌트 분리 및 포인트 상세 페이지 추가
