+++ src/app/fishing-point/components/BannerSection.tsx
