+++ src/app/fishing-point/components/FishingSpotsSection.tsx
